requirements.txt
python-dotenv==1.0.1
tqdm>=4.66.4,<5.0.0
transformers>=4.41.2,<5.0.0
torch==2.3.1
huggingface-hub>=0.23.2,<0.24.0
requests>=2.32.3,<3.0.0
python-magic>=0.4.27,<0.5.0
langchain>=0.2.2,<0.3.0
aiohttp==3.9.5
